requirements.txt
scikit-learn
pandas
numpy
tqdm
timm
wandb
h5py
pytorch-metric-learning
augly
pillow
faiss-gpu
matplotlib
albumentations
opencv-python-headless
ptflops
torch_optimizer
seaborn
